+++ CHANGELOG.md
@@ -18,6 +18,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial state styles
 - Searching styles
 - State handling using new react context
+- Pickup points are now searched with checkout API simulation
+- External pickup points are now searched with the new pickup points endpoint
 
 ## [2.3.11] - 2019-05-21